finished

+++ app/Http/Controllers/CategoryController.php
@@ -5,16 +5,25 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Category;
+use App\Models\Post;
+use Illuminate\Support\Facades\DB;
+
+
 
 class CategoryController extends Controller
 
 {
     public function getCategories()
     {
-        $categories=Category::first()->posts;
-       // dd(Category::first()->posts()->get());
-       return response()->json(['categories'=>Category::first()->posts()->get()],200);
+       // $categories=Category::first()->posts;
+        
+       //dd(Category::first()->posts()->get());
+       return response()->json(['categories'=>Category::all()],200);
+
     }
+    
+
+      
     public function createCategory()
     {
         $validation=Validator::make(request()->all(),[
@@ -48,8 +57,12 @@ class CategoryController extends Controller
     public function deletecategory($categoryId)
     {
         $category=Category::find($categoryId);
+        
         if(!$category){
             return response("could not finde category",404); 
+        }
+        if($category->posts->count()>0){
+            return response("cant delete this cat",404); 
         }
         $category->delete();
         return response("deleted",200);
